Add follow-data animation
This adds a new animation mode which will track a data object
as it moves during the animation. The cameras position and focal
point will be updated each time-step to keep the data at the center
of the view.

In order to make this change the vtkPVCameraCueManipulator class
now needs access to the vtkSMSourceProxy for the data to follow
so that it may get the bounds from the data information. This
necessitated moving the animation classes from ClientServerCore
to the ServerManager module.

Change-Id: I9d82f749d459bb7598a96f77dd5dbc049e9bfce3