BUG: Avoid redistribution of stuctured grid when volume rendering.
Due to a bug in logic, despite best intentions
vtkStructuredGridVolumeRepresentation was ending up marking its data as
redistributable. Consequently, vtkPVDataDeliveryManager was redistributing it
which was negating the whole purpose of having
vtkStructuredGridVolumeRepresentation. This fixes that issues by adding API to
vtkPVRenderView to call MarkAsRedistributable() to unmark the representation as
redistributable.

Change-Id: I8474d2f9246b90b3989510688e2a6d67223f5fcb