Fix transfer function rescale in log-scale mode.
When loading color-maps or rescaling transfer function points when log-scale was
enabled, we do all the necessary interporlations in log-space. ApplyColorMap(),
InvertTransferFunction(), and RescaleTransferFunction()  now respect log-scale
mode.

Also fixed entering in and out of log-space. Now when user click "use-log-scale"
check box, the control points are respaced in the requested space so that the
transfer function appears the same (though the mapping changes). Added a new
test (LogColorMap) to test for log-color maps.

Change-Id: I67f13abd394cdb2fbe2a710b51f209f066291ab0