BUG #14187. Fix default to not cache animations.
ParaView by default caches animation geometry. With cached geomtry the
data-ranges cannot be shown correctly. Since geometry-caching happens under the
covers it difficult for users to understand what's going on with data-ranges.
Hence, we change the default to not cache. Thus when users do turn on
geometry-caching, they can be made away that that will affect data ranges.

Change-Id: I4445fb0f9c71f6701edd1db3bbcb779c75419adf