Use Python wrappings to implement ClientServer wrappings
With the unified bindings option set, ParaView will use Python as the
backing implementation for ClientServer wrapping. The goal is to
minimize the size of the build for Catalyst and the CS wrapping code is
an obvious target.

ParaView's plugins have unified bindings forced off because they cause a
circular dependency which cannot be broken (the plugin library contains
the ClientServer wrappers which, with unified bindings, requires the
Python library which requires the main plugin library for the actual
work).

Change-Id: Ieeb115799147c44822831f9c569fa37c4146b891