Changing the writers to be available on the client based on server XML.
Same changes that I just did for readers to make them only defined
on the server and then pass that info to the client. Also removed
the GMVReader plugin's GUI XML that isn't needed. Finally, added
in warnings for both reader and writer plugins that still define
the GUI XML stuff instead of putting the hints in the server
side XML Hints section.

Change-Id: Idbf592f0cebcb69a214e1677743cf64318110380