Add context to new instance and command functions
This allows other mechanisms than just C functions to be registered with
the ClientServerInterpreter. Particularly, it will allow for using the
Python wrappings if available rather than building CS wrappings as well.

The implementation would use std::function if supporting it from MSVC
2008 through more modern compilers were easier (MSVC doesn't use a tr1/
directory for TR1 includes and not all TR1 functionality is available as
std::tr1 (just std::)).

All command function calling is internalized as part of this since the
context is only known to the vtkClientServerInterpreter class. In
particular, the GetCommandFunction method is replaced with a
HasCommandFunction/CallCommandFunction method pair.

Change-Id: I827ccc42e69f676e48a8453123834d6dbbb535a5