BUG #14158. Making it easier to move sliders when they overlap.
When the min and max sliders overlap, it's very hard for the users to move them.
To make that easier, added new ability to double click on the empty space on the
slider widget to move the closest slider to that position. Thus, when the two
sliders overlap, the user can double click on the left-side of the two sliders
to move the min or right-side for the max slider.

Change-Id: I24073cf403661434302ceaf2d7d0caa784f999e6