BUG #14201: Fix default chosen for x-axis.
Addresses BUG #14201 by changing the logic used to pick the default for X-axis.
The bug was that for histograms, we were choosing the wrong array as the x-axis.
We should move this logic to a domain that can reside within the server-manager.
But we leave that for a later stage when we cleanup the chart representations.

Change-Id: Ib8bd9a75fa6206d7ae8cadab5f07113d87b7e15c