BUG #14218. Fixes color preset issues.
The color preset dialog showed up with the first item active. As a result when
the user clicked on it, it wouldn't realize that the user selected and it and
hence would never change to that item causing the reported bug. We fix it by
using the selection-changed signal rather than the current-changed to catch this
case.

Change-Id: I1e99fc25cfc8095bef80afdf888dcf36b7d6b846