Refactor the Find Data dialog
This refactors the Find Data dialog and incoroprates into it the
functionality of the now-deprecated selection inspector panel.

The Find Data dialog now allows for users to both create new selections
and view the current global selection. Also allows the user to update
display properties (e.g. color, font) for the selection labels.

Change-Id: I4bac1cfe64e47b7c9cf7d14b9a8021a8b6d59650