BUG #14253. Fix translucent slice representations rendering.
When rendering image-slices with translucency, vtkImageSliceRepresentation was
not telling the view to use ordered compositing. That was causing the rendering
to fail as manifested by the bug #14253. This fixes that issue.
vtkImageSliceRepresentation now correctly requests ordered compositing when
needed.

Also fixed a bug in vtkOrderedCompositeDistributor which resulted in the filter
crashing when a process ended up with an empty dataset after redistribution for
ordered compositing. We just needed to check if the data has valid points before
we attempted to extract the exterior surface from it.

Change-Id: I2b1483e98c8a9c95be9e3076252badc6217cf0ab