Fix paraview's python modules.
This change fixes the install rule for paraview's python modules.
It also copies copies vtk's python modules into the build tree in two
locations. The first is into paraview/vtk/, for old style scripts that
use a subset of vtk. The second is into vtk/ for new scripts,
that want to access the whole of VTK.

I've modified the module search path for linux so that we can
have one python module install location that works for both VTK and
ParaView under Mac, Linux and Windows.

Change-Id: I2e81232502c8905bf3fdd1d4aaec5e4cc6e5aa7e