Enhancements to particle streaming.
Enhancements to particle streaming to allow purging of particles as user zooms
out -- consistent with past versions of the functionality. Also added mcehanism
to show outlines for blocks being fetched as well as change opacity for the
particles.

Updating vtkStreamingParticlesPriorityQueue to make it easier to change the
comparison operation based on the nature of the dataset.

Change-Id: I96d2bee8cc3b4782496a3ecae2f07411a53fa7f0