GenericIOReader cleanup & fix CosmoTools warnings
This commit cleans up some of the code in the GenericIO reader
such that it is more modular. All internal
helper methods have been moved to a separate header,
vtkGenericIOUtilities.h, to make the code cleaner.
Further, the properties associated with querying rank
neighbors (which can be used when reading data in RCB)
have been marked as advanced. Last, a number of compiler
warnings have been fixed.

Change-Id: I8e66b266c89e3dc1e2ae9c4ef0ab8e63893429fc