Added mechanism to link properties in representations.
In a typical use-case for Representation proxies, we have properties on the
outer representation proxy, e.g. Input, which is also present on sub-proxies.
Now, the Input property on the sub-proxies needs to be linked with the outer
"Input" property so that domains are updated correctly when the use changes the
outer "Input" property. This was done by a hack in vtkSMPVRepresentationProxy.
Since this pattern is typical with representations, providing a mechanism to
setup links generically by adding an API to vtkSMRepresentationProxy.

Also fixed vtkSMProperty subclasses to treat vtkSMProperty::Copy() to imply that
both checked and unchecked property values are copied. The newly added API to
vtkSMRepresentationProxy relies on that.

Change-Id: Iab5fa11c92032ebee6b9ffbe84172d8cec41e32c