BUG #14255: Fixed issues with scalar-translucent rendering in parallel.
Due to the logic to skip depedency tracking of LUTs and piece-wise functions,
when one swicth the LUT to enable opacity mapping for surfaces, the view wasn't
being marked dirty and hence didn't call "update" on all representation thus
realizing that ordered compositing is needed. We had added the logic to skip
depedency since we didn't want the geometry to be invalidated when LUT changes.

Hence, added a hackish stub to not skip depedencies for LUTs but to skip
marking the representation dirty for them explicitly.

Change-Id: I21afa312c94e320231870a674e406b11b56fc46d