BUG #14289. Fix extent issues with programmable-filter.
vtkPythonProgrammableFilter created a OnePieceExtentTranslator by default. That
resulted in all processes requesting the whole-extent on each of them which
resulted in BUG #14289. This was done so to handle a use-case where we used
vtkPythonProgrammableFilter to generate structure-data for operations like
histogram or stats. VTK has since then moved to using vtkTable for such
operations hence the need for OnePieceExtentTranslator becomes moot. While,
users using vtkPythonProgrammableFilter without expecting to change the process
extents when dealing with partitioned image data have strated to grow. Hence
changing the default behavior of the filter to be as expected -- i.e. not muck
with data extents -- in parallel.

Change-Id: I2e8d695bd369d1a7ba4b6b5d64eb3ecb03440e6a