BUG #14290: Fix static build issues with plugins and no-gui.
When PARAVIEW_BUILD_QT_GUI was Off, and certain plugins that created GUI-only
plugins were enabled, we'd get build errors. This was due to the fact that we
were not realizing that those plugins, despite being enabled were not adding any
plugin library to be built. This fixes that issue by validating all plugin
library names are plugins are being processed.

Change-Id: I618cf7065bd785c27381cfb7ff2cff8943cac919