BUG #13797. Fixed chart view size issues with pvbatch.
Chart view sizes weren't having any effect in pvbatch. The problem was that the
logic to resize the windows which was in vtkPVSynchronizedRenderWindows was
never  called since the vtkPVSynchronizedRenderWindows instance is not enabled
for chart views (since they don't do parallel rendering). It worked for 3D views
since vtkPVRenderView always enabled the vtkPVSynchronizedRenderWindows in batch
mode.

Fixed the logic by making sure that vtkPVSynchronizedRenderWindows lays out the
views/renders in batch mode even for chart views. This also address the bug fix
for BUG #14280.

Change-Id: I3c99a7a267bbe8ad45b7fbc428b3e13ed3433e43