Fix CoProcessingTestPythonScript test.
CoProcessingTestPythonScript fails on Windows because when the Python module
libraries are loaded, they cannot find the dlls for the Python modules since
they are in bin. The test runs with a different working directory than the bin
directory. Hence we fix by ensuring that on Windows, the PATH is extended to
include in ParaView build dir's bin directory when setting up the Python
environment.

Change-Id: I92c29544114a567bbf62e4553c6c76aa800c9ba4