BUG #14320: Turn on user-prompts on error on Windows.
In general turn off error prompts. This is where the process waits for
user-input on any error/warning. In past, we turned on prompts on Windows.
However, for client processes, we capture the error messages in the UI and
for the server-processes, prompts can cause unnecessary interruptions. Hence
we turn off prompts all together.

Change-Id: Idc8b27485ab1d6a7b70580a38d9b4d7340688e0d