BUG #14186: Fix issues with scaling for point sprites.
The problem was arising from the fact that the properties for selecting array to
scale radius or opacity by were being set to default values prematurely. Due to
the extensive custom logic in the point sprites code, setting the array names
was not sufficient and caused inconsistencies between what was shown in the UI
and what was happening under the covers. Fixed the issue by ensure that by
default, we don't scale my any array.

Also fixed issue with loading state, where the radius range was getting reset
when the panel was created for a visualization loaded from state files.

Change-Id: Icc19ad145cafb2cfed59e88c617e931ab303b72c