BUG #14343: Fix hang in tile-display mode.
When in data-server/render-server mode with tile-display, ParaView was hanging
when rendering data. This was due to the fact that the data-server never
realized that it was in tile-display mode since the -tdx/tdy parameter is
specified on the render-server and was never communicated to the data-server.
Fixed that issue by adding a communication step between data-server and
render-server during initialization to ensure that they are on the same page as
far as the tile-display mode goes.

Change-Id: If1690293c4087adff5ffa1affd4103a053a1cbde