Added a new ParaViewWeb application to be used only for testing.
Added a new python server and a very basic index.html for a new
ParaViewWeb application which will be used only to test some new
module registration functions and to test the various paraview
web protocols.  Also added some bits to the web applications
CMakeLists.txt which will copy the new app to the build tree only
when testing is enabled.  Modified this same file to recursively
copy the new test folders inside any web application to the build
tree, but again, only when testing is enabled.

Change-Id: I65dbcd2be7bd6a0ffbc5c496069ebacb26046db4