Added a CMake macro to find the selenium drivers on a system.
Added a ParaView test macro for adding web-based tests.

This macro takes as arguments a list of browsers that are required by
the set of test scripts, as well as a list of the required Python
modules.  The macro checks that the selenium browser drivers are
actually present, and that the required Python modules are present as
well, before adding web-based tests for each browser (adding each test
script with every browser).

Change-Id: Ieaa7f918303d2dfd777e7ed98ffe3c9d41038845