BUG #14287: Fix find-data not updating with time changes.
Ensure that pqFindDataCurrentSelectionFrame updates the spreadsheet whenever the
data pipeline is updated for the seleceted port.

Also fixed logic in vtkSMProxy which was resulting in duplicate firing of
vtkCommand::UpdateDataEvent event if the UpdateDataEvent callback resulted in
calls to request further updates to the pipeline e.g. render some other view.

Change-Id: I51469a0e81a64a7bd314c90a8c9e0ee97312e32b