Provide a ParaViewWeb standalone test application.
Added a new ParaViewWeb application which allows to register
individual Javascript functions which test arbitrary parts of
ParaView or VTK Web.  The application provides a simple user
interface which allows selecting tests to run, then viewing
the status of each of the tests.

Also updated a Python/selenium test which can run against this
application, initiate the Javascript tests with a button click,
then wait for all tests to finish.  If there were any failures,
they are reported back to ctest.

Change-Id: Id636b47b73a9b72bd0c12df2f13e828eded15104