Fixed some broken tests of the new ParaViewWeb test application.
One broken test was the python test which retrieved the index.html
and verified presence of script includes.  Now this test actually
attempts to load the needed javascript files and checks the HTTP
return code.

The other broken test was due to assuming known indices of data
items in the ParaViewData repository.

Change-Id: Idf22d90f48c12146850f5ec89d388d9552dc191c