Remove reference counts from vtkSMDataTypeDomain
With unified bindings enabled, this cache's pointers get mixed up with
the Python object cache's pointers resulting in double free problems
since both store the pointer with vtkSmartPointer. The reference
counting seems to aim to allow dropping the cache after it isn't
"needed" anymore, but with Python objects floating around, at least one
instance is probably in the cache, pinning the reference count to
non-zero, so it likely has little effect in a unified-bindings world.

Change-Id: Id45ab6c11acf257f1ec989acd83f4f7c2c6d52a1