Added advanced color and opacity map controls
These controls are tables that display the scalar values and assigned
colors and opacities for the current color and opacity maps,
respectively. Colors and opacities are limited to the range [0, 1].
Colors are specified as RGB triplets. The first and last scalar values
cannot be modified through the table as these are set via different
controls.

In addition, tabular formatted text data can be pasted from the
clipboard using Ctrl-V. The tabular text data must be delimited with
an arbitrary amount of whitespace (no newlines allowed). Pasting
always begins from the upper left to the bottom right table cells.

Change-Id: I2c10c9306d7e98c45cc37cec2ac16683876182f9