BUG #14451: Restore connect-id authentication for client-server.
Restoring logic to validate connect id when setting up client-server
connections. The connect-id is now included in the handshake message which is
used to validate the version compatibility.

Also fixed assertion failure in pqCommandLineOptionsBehavior when a url was
specified on command line, but the server rejected the connection due to
mismatched handshake.

Change-Id: I752d4237e15533e13db8ba12788dd966ef9b26d3