BUG #14286: Fixed orientation widget now showing up in Catalyst.
Orientation widget was hidden in pvbatch or catalyst since the renderer for the
orientation widget was never been unhidden by vtkPVSynchronizedRenderWindows.
Fixed that by adding logic to vtkPVRenderView to register the orientation axes
rendendere with vtkPVSynchronizedRenderWindows. Added logic to
vtkPVSynchronizedRenderWindows to handle this renderer correctly. In future we
may fix vtkPVAxesWidget to not use the renderer all. Until then, we live with
this approach.

Change-Id: Ide8196bada06a95564f7ecddb04e9971f93bc650