Added busy state event to pipeline js, also fired addSource when file loaded.
The new busy status event is fired with an argument of true before each rpc
call, and then fired again with false when the rpc call finishes, whether
the call failed or finished successfully.  This will allow widgets to
know about the busy state of the pipeline and respond accordingly.  Also
added code to fire a pipeline addSource event when a data file has finished
loading.

Change-Id: I3fb1e7fee84e20bd3f5ad443b5a5d8996917fd24