BUG #14214: Miscellaneous fixes for tracing.
Cleaned up logic in tracing. We were ending up with out of order trace captures
resulting in broken trace files that didn't work as expected using all ParaView
executables. This was happening because a lot of new code added to smtrace
directly appended to the trace_output stream without flushing the information
currently gathered. Fixed that by streamlining the logic.

Also fixed a leak with trace observer. The observer was in place even after
tracing was stopped. Fixed that.

Change-Id: I805ffd7570c0b79ace9b540504a55c8f5e842b23