Added new protocols to dynamically load plugins and filters.
Added two new paraviewweb protocols.  One allows to specify a list
of plugins to be loaded upon startup of the web application.  The
other allows to specify a list of filters in a json text file which
will be available in the web application.  This is an improvement
on the static filter list previously provided by the pipeline
manager protocol.  The web visualizer application has been updated
to give examples of how these two protocols are used.

Change-Id: I0d2f556347a777a04c82c7b04613ee6c8a3d2f9a