Add new serverReady signal for proper collaboration initialization
Previously the PluginManager was listening to preServerAdded
to load plugins from user settings. And when a plugin was loaded
it was triggering some client/server communication for synchronizing
the proxy definition while the ServerManager was not ready yet.

In collaboration mode, this was forcing the ServerManager to push its state
to the server before getting the proper state from the server.
Hence each new connection was reseting the full shared proxy manager state.

This was introduced while trying to fix bug #12238.

Change-Id: Icb6c7e39a28e150095e6c77b6417e868657f5c05