Allow a title bar to be resized smaller than its width.
If a title bar becomes smaller than its width (the sum of widths of all its
icons) use an extension button to show hidden buttons.
Note that the title bar is as small as before but there is some space
between the title bar and the extension menu (when that appears).
It seems this is because the tool bar wants to be a certain size.

Change-Id: I53a019505937462c8aba20edb5c17df720a19e9e