Cleanup trace generated for Exodus reader.
Exodus reader was generating unnecessary entries in the trace file. This was
because the XML was setting a default number of values count for server
properties which wasnt needed. Fixed that.

Also fixed vtkSMVectorPropertyTemplate to fire modified event in SetElements(..)
for uninitialized properties only when the new value had a non-zero number of
elements.

Change-Id: Iab759d282e80e7818c67dbf8a002e8a37167b729