Fix problem reading buffer from matplotlib
matplotlib.backends.backend_agg.FigureCanvasAgg.print_to_buffer()
prints an image to an old-style Python buffer object and ParaView was
using the API for the new-style Python buffer object. Serendipitously,
this works in Python 2.7.5 (on my development machine), but that is
not so on other Python versions. Switched to using the API for the
old-style buffer object and verified that it works with Python 2.7.2.

Change-Id: Idfa91c27109756e275d690229865e5f648249bda