Cleanup logic for setting default values.
Added mechanism on vtkSMProxy to set default values for all properties that
pqProxy now uses when a new proxy is created. We still have a few hacks lurking
around for dealing with compound proxies in pqPipelineSource.

Added support in vtkSMChartRepresentationProxy to correctly set UseIndexForXAxis
property's default value based on what XArray is picked.

Change-Id: If7e4ba6d40541e501a5524d2bbe55be8f992d646