Fix for space left between the title bar and extension menu.
If a title bar is smaller than its width an extension menu shows
the hidden buttons. Because one of the buttons uses text and
the title bar is resized smaller than what the font allows,
there is some space left between the title bar and the extension menu.
We replaced the text with an icon which fixed this problem.

Change-Id: I02fac2f67971581605a45b7d87a4f1de761816c7