Query the 'name' attribute if 'function' is not defined for a Property.
When defining a pqPropertyWidget subclass for a PropertyGroup,
the 'function' attribute of the Property element is used to
specify the functionality queried by the
custom widget. This allows to use the same custom widget for different
property groups. However this also forces one to define the 'function' attribute even if the custom widget is only applied to one property group. For those cases, we use the 'name' attribute if 'function' is not defined.

Change-Id: I57b75c4fcfe3201eb615f29e258303f7a7937335