BUG #14496. Fixes issues with tracing certain proxy properties.
When on changed helper property values e.g. the Normal for a clip plane, trace
would not record it unless the creation the "Clip" filter was also captured in
that trace. That was due to missing logic to "locate" the proxy being modified
when the Normal was changed. This patch addresses that issue.

Change-Id: I9738ed98cdd19671a708f1c8f72057bc4e43628b