BUG #14512: Print debug info when connection fails.
If a client-server connection handshake fails, it can be tricky for users to
determine exactly what went wrong. To make it easier, each process now prints
out information on the standard output providing information about that process
such as the handshake used by that process, the socket communicator vesion
number. The user can then compare the values printed out on the client and
server side to deduce why the connection failed.

Change-Id: Ib7f2adf477abf2cc865f55870dbf87de7557f029