BUG #14535: Adding logic to pop out views in a separate window.
Users can now pop-out a "layout" form the MainWindow into a separate window. A
new icon on the right side of the layout text can be used to pop out the layout.
Clicking on the icon again or simply closing the popped out window will put the
layout back into the main window.

There's one outstanding usability issue with this enhancement: shortcuts (such
as 'S' for selection) don't work in the popped out window. The MainWindow needs
to have the focus for that to work currently.

Change-Id: I93139faffffd1190628ab43c1fc7a1d512cdf06f