Test to verify that ParaView handles a plugin loaded multipe times.
this test is not specific to the EyeDomeLightingView plugin, it is meant
to test that we can reload a plugin multiple times without a crash.
for complex Python scripts we may not know that a plugin has already
been loaded and paraview should not crash if it does. this was the
behaviour before fixing the bug in the previous commit too.

Change-Id: I8f8c9809bf55a2b9d1ada8c23ab7748291e16757