Added vtkSMSettings class
This class is meant to be a central location for user-defined program
defaults. It supports the notion of site-wide settings meant to be
used for installations of ParaView in a centralized location for a
specific computing site. Site-wide settings can be overridden by
user-specific settings specified in a file in the user's home
directory.

The site-wide file is named .pvsitesettings.js and its location has
yet to be determined.

The user-specific a file is named .pvsettings.js and it is expected to
be in the user's home directory. An example file for changing the
properties in the SphereSource looks like:

{
  "sources" : {
    "SphereSource" : {
      "Radius" : 4.0,
      "ThetaResolution" : 32,
      "PhiResolution" : 32,
      "Center" : [1.0, 2.0, 3.0]
    }
  }
}

A common pattern of usage is expected to involve calling

vtkSMSettings::SetProxySettings(vtkSMProxy* proxy)

after a proxy is created in the application.

Change-Id: Ieaca8f4d1ea7258372bf84af87b6a1bae0c24ff5