Cleanup chart axis range reset logic.
Every time the view is updated, we recalculate the chart axis ranges. The
recalculation logic respects the case when user has "fixed" the axis range so
it's not a problem to do this on every update. This ensures that over time and
when new series are added/removed, the axis range is update automatically.

Change-Id: I4a3a706ee391d62177f9211c7371325d580e6365