BUG #13773. Fixed auto-apply issues with plot-selection-over-time.
pqSelectionInputWidget was setting up the default value for the selection input
using a timer. That was happening after the auto-apply timer timed out. Thus,
the apply would happen before the value was set, resulting in the default for
the representation being setup incorrectly. That meant no data would be shown
even the selection-input was setup immediately after when the timer expired.
This fixes that issue by ensure that pqProxyPropertyWidget explicitly does the
initialization and removing the timer logic from pqSelectionInputWidget.

Also added a test to test this.

Change-Id: I6b51801e362b0cbc9aec4a7a370db0b6f32adcb2