BUG: pqFileDialog adds a separator even if only one file selected
The separator adding logic was looking at number of indices selected which was
wrong. When a file is selected, there can be multiple indices based on the
number of columns associated with it. Looking at the number of rows selected in
this case gives the correct number of files selected.

Fixes issue 14455 (http://www.paraview.org/Bug/view.php?id=14455)

Change-Id: I82ddfe6a350d27f90d56e8ba7629ef5e06ac1c34