Working on cleaning pqDisplayColorWidget.
Making pqDisplayColorWidget use the domains instead of
pqPipelineRepresentation. Moving pqDisplayArrayWidget back to
PointSprite plugin. It was moved into ParaView for no particular reason.

Also fixed issue with ColorArrayName's domain not updated property when
inital representation is Outline.
When creating a Wavelet source, for example, the domain would not update
to include cellNormals array even after we switch representation type to
"Surface" from "Outline". This was because the SurfaceRepresentation
proxy's data information was never invalidated when representation type
was changed. It makes sense to invalidate subproxies when properties on
the parent-proxy are modified. This fixes the cellNormals issue too.

Change-Id: I95aae3907dbae24a011fbc9f7a15d427da514245