Fixed logic to change representation type.
When representation type is changed, in some cases we are needed to
setup scalar coloring and other things to ensure that the
representation type can function e.g. when picking Volume or Slice
representation. Added mechanism in vtkSMPVRepresentationProxy to
encapsulate the logic for these checks. pqDisplayRepresentationWidget
now uses the vtkSMRepresentationProxy::SetRepresentationType() method to
change representation type, rather than just changing the property.

Change-Id: Ib6be291208bb0c505624986a00843be1fe47acee