Imbued vtkSMProperty with a vtkSMPropertyLink instance
This enables other properties to take on the value of a linked
property whenver the value of the linked property changes.  Added
method AddLinkedProperty(vtkSMProperty* property) to vtkSMProperty to
make linking simple. An arbitrary number of properties may listen
for updates from a linked property.

With this mechanism, links between settings proxies can be set up
easily with some XML hints. Added these hints to the appropriate XML
files.

Change-Id: I1a39924eee05d2779898fef0fc5f44635b706aca