Target properties in a property link can now break the link
When a target property that listens for updates from a source property
is deleted, it can now remove itself from the source property's list
of linked objects. This is done in the property destructor to ensure
the source property does not have a link to a NULL object.

Removed some linked property list cleaning from vtkSMPropertyLink that
was kind of a hacky way to deal with this before.

Change-Id: I510513e0e17bef5357e5e8c935a593c10b0bba0e