Quick-n-dirty pass at fixing the depth buffer in parallel.
Here's a first pass at fixing the depth buffer in parallel for
vtkPVRenderViewForAssembly. To further clean this up, we should do the
following:

+ Cleanup API on vtkPVSynchronizedRenderer to provide access to depth
buffer and color buffer
+ Teach vtkPVSynchronizedRenderer to user color and depth buffer from
IceT when available and grab from frame-buffer when not.
+ vtkPVRenderViewForAssembly should then use that API to get the color
and depth buffer (never having to do actual grabs).

This commit exposes internal methods to use for
vtkPVRenderViewForAssembly as a quick hack. We can come back and fix the
API in future.

Change-Id: I4c62e94377750202a43ac6682feb2d72bf17da58